Allow locating errors outside of isomorphisms (#155)
We now locate the error message based on line number and not just
tactic feedback.  This handles cases where the LLM has removed the
initial `iso.` part of the proof, and also cases where the LLM has added
a lemma and the issue occurs in that lemma.  Correspondingly, we now
allow the LLM to remove lemma blocks and also to edit/replace lemma
blocks.

Fixes #154